# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "TerrAPI is a service designed to simplify and automate your deployments.",
        "title": "TerrAPI Controller",
        "contact": {
            "name": "Support",
            "url": "https://github.com/terrapi-solution",
            "email": "contact@thomas-illiet.fr"
        },
        "version": "1.0"
    },
    "paths": {
...
      "/v1/activities": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                  "🍍 Activity"
                ],
              "summary": "List all activity.",
                "parameters": [
                    {
                      "type": "string",
                      "description": "Search",
                      "name": "search",
                      "in": "query"
                    },
                    {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "collectionFormat": "csv",
                      "description": "Filter",
                      "name": "filter",
                      "in": "query"
                    },
                  {
                    "minimum": 1,
                        "type": "integer",
                        "default": 1,
                    "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 10,
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    },
                  {
                    "type": "string",
                    "default": "id",
                    "description": "Order by",
                    "name": "order_by",
                    "in": "query"
                  },
                  {
                    "type": "string",
                    "default": "desc",
                    "description": "Order direction",
                    "name": "order_direction",
                    "in": "query"
                  }
                ],
              "responses": {
                "200": {
                  "description": "OK",
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/definitions/activities.ActivityResponse"
                    }
                  }
                },
                "500": {
                  "description": "Internal Server Error",
                  "schema": {
                    "$ref": "#/definitions/errors.AppError"
                  }
                }
              }
            }
      },
      "/v1/activities/{id}": {
        "get": {
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "🍍 Activity"
          ],
          "summary": "Get a specific activity.",
          "parameters": [
            {
              "type": "integer",
              "description": "Activity identifier",
              "name": "id",
              "in": "path",
              "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                          "$ref": "#/definitions/activities.ActivityResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                          "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                          "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
      "activities.ActivityResponse": {
            "type": "object",
            "properties": {
              "createdAt": {
                    "type": "string"
                },
              "deploymentId": {
                "type": "integer"
              },
                "id": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
              "type": {
                    "type": "string"
                }
            }
        },
...
      "errors.AppError": {
        "type": "object",
        "properties": {
          "err": {},
          "type": {
            "type": "string"
          }
        }
      },
...
    },
    "securityDefinitions": {
        "JWT": {
            "type": "apiKey",
            "name": "token",
            "in": "header"
        }
    }
}